+++ src/byob.js
@@ -113,7 +113,7 @@ ADT_SlotMorph*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.byob = '2026-March-09';
+modules.byob = '2026-March-10';
 
 // Declarations
 
@@ -4169,7 +4169,7 @@ BlockLabelFragment.prototype.defTemplateSpecFragment = function () {
             this.defaultValue ? ' = ' + this.defaultValue.toString() : ''
         );
     } else if (contains(
-            ['%cmdRing', '%repRing', '%predRing', '%anyUE', '%boolUE'],
+            ['%cmdRing', '%repRing', '%predRing', '%anyUE', '%boolUE', '%nUE'],
             this.type
         )) {
         suff = ' \u03BB';
@@ -4900,8 +4900,9 @@ InputSlotDialogMorph.prototype.createSlotTypeButtons = function () {
     defLabel.refresh = () => {
         if (this.isExpanded && contains(
                 [
-                    '%s', '%n', '%txt', '%anyUE', '%b', '%boolUE', '%adt',
-                    '%mlt', '%code', '%upvar', '%parameter', '%clr', '%mult%adt'
+                    '%s', '%n', '%nUE', '%txt', '%anyUE', '%b', '%boolUE',
+                    '%adt', '%mlt', '%code', '%upvar', '%parameter', '%clr',
+                    '%mult%adt'
                 ],
                 this.fragment.type
             )) {
@@ -4929,12 +4930,12 @@ InputSlotDialogMorph.prototype.createSlotTypeButtons = function () {
     defInput.setWidth(50);
     defInput.refresh = () => {
         if (this.isExpanded && contains(
-            ['%s', '%n', '%txt', '%anyUE', '%mlt', '%code', '%upvar',
+            ['%s', '%n', '%nUE', '%txt', '%anyUE', '%mlt', '%code', '%upvar',
                 '%parameter', '%adt', '%mult%adt'],
             this.fragment.type
         )) {
             defInput.show();
-            if (this.fragment.type === '%n') {
+            if (['%n', '%nUE'].includes(this.fragment.type)) {
                 defInput.setIsNumeric(true);
             } else {
                 defInput.setIsNumeric(false);
@@ -5316,7 +5317,7 @@ InputSlotDialogMorph.prototype.addSlotsMenu = function () {
             (contains(
                 ['%mlt', '%code', '%obj', '%scriptVars', '%receive', '%send',
                     '%elseif', '%upvar', '%mult%upvar', '%parameter', '%adt',
-                    '%mult%parameter', '%mult%adt'],
+                    '%mult%parameter', '%mult%adt', '%nUE'],
                 this.fragment.type
             ) ? on : off) +
             localize('special'),
@@ -5367,6 +5368,7 @@ InputSlotDialogMorph.prototype.specialSlotsMenu = function () {
 
     addSpecialSlotType('multi-line', '%mlt');
     addSpecialSlotType('code', '%code');
+    addSpecialSlotType('number \u03BB', '%nUE');
     addSpecialSlotType('object', '%obj');
     addSpecialSlotType('parameter', '%parameter');
     addSpecialSlotType('ADT', '%adt');